~Fixes:  - added directory fixtures for fixtures keeping  - changed custom command, now contains 2 functions delete and load from fixture
